fix: color to recipe
