+++ src/features/common/components/Typography/type.ts
@@ -1,0 +1,1 @@
+export type TxtColor = 'primary' | 'textPrimary' | 'textSecondary' | 'inherit'
